+++ navigation/index.tsx
@@ -14,24 +14,21 @@ import Colors from '../constants/Colors';
 import useColorScheme from '../hooks/useColorScheme';
 import ModalScreen from '../screens/ModalScreen';
 import NotFoundScreen from '../screens/NotFoundScreen';
-import HomeScreen from '../screens/HomeScreen';
-import PuchiPuchi from '../screens/Puchipuchi';
-import Gyro from '../screens/Gyro';
+import HomeScreen from '../screens/home/HomeScreen';
+import PuchiPuchi from '../screens/solutions/puchi_puchi/PuchipuchiScreen';
 
-import AddStress from '../screens/AddStress';
-import Batting from '../screens/Batting';
-import CutAction from '../screens/CutAction';
-import History from '../screens/History';
-import Jogging from '../screens/Jogging';
-import SelectDeadline from '../screens/SelectDeadline';
-import SelectSize from '../screens/SelectSize';
-import SolutionSelect from '../screens/SolutionSelect';
-import StressSelect from '../screens/StressSelect';
+import AddStress from '../screens/add/AddScreen';
+import Batting from '../screens/solutions/baseball/Batting';
+import CutAction from '../screens/solutions/cut/CutScreen';
+import History from '../screens/history/HistoryScreen';
+import Jogging from '../screens/solutions/jogging/JoggingScreen';
+import SolutionSelect from '../screens/solutions/SolutionScreen';
+import StressSelect from '../screens/stress_list/StressListScreen';
 import WriteStress from '../screens/WriteStress';
+import SettingsScreen from '../screens/SettingsScreen';
 
 import { RootStackParamList, RootTabParamList, RootTabScreenProps } from '../types';
 import LinkingConfiguration from './LinkingConfiguration';
-import SettingsScreen from '../screens/SettingScreen';
 
 export default function Navigation({ colorScheme }: { colorScheme: ColorSchemeName }) {
   return (
@@ -57,15 +54,12 @@ function RootNavigator() {
       <Stack.Group screenOptions={{ presentation: 'modal' }}>
         <Stack.Screen name="Modal" component={ModalScreen} />
         <Stack.Screen name="PuchiPuchi" component={PuchiPuchi} />
-        <Stack.Screen name="Gyro" component={Gyro} />
 
         <Stack.Screen name="AddStress" component={AddStress} />
         <Stack.Screen name="Batting" component={Batting} />
         <Stack.Screen name="CutAction" component={CutAction} />
         <Stack.Screen name="History" component={History} />
         <Stack.Screen name="Jogging" component={Jogging} />
-        <Stack.Screen name="SelectDeadline" component={SelectDeadline} />
-        <Stack.Screen name="SelectSize" component={SelectSize} />
         <Stack.Screen name="StressSelect" component={StressSelect} />
         <Stack.Screen name="SolutionSelect" component={SolutionSelect} />
         <Stack.Screen name="WriteStress" component={WriteStress} />
@@ -84,7 +78,7 @@ function BottomTabNavigator() {
   const colorScheme = useColorScheme();
 
   return (
-    <BottomTab.Navigator
+        <BottomTab.Navigator
       initialRouteName="Home"
       screenOptions={{
         tabBarActiveTintColor: Colors[colorScheme].tint,
