+++ package.json
@@ -24,7 +24,9 @@
     "expo-asset": "~8.7.0",
     "expo-constants": "~14.0.2",
     "expo-font": "~11.0.1",
+    "expo-haptics": "~12.0.1",
     "expo-linking": "~3.3.0",
+    "expo-sensors": "~12.0.1",
     "expo-splash-screen": "~0.17.5",
     "expo-status-bar": "~1.4.2",
     "expo-system-ui": "~2.0.1",
@@ -33,12 +35,11 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-flex-layout": "^0.1.5",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
     "react-native-web": "~0.18.9",
-    "typescript": "^4.6.3",
-    "expo-sensors": "~12.0.1",
-    "expo-haptics": "~12.0.1"
+    "typescript": "^4.6.3"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
