+++ screens/stress_list/StressListScreen.tsx
@@ -1,15 +1,15 @@
 import { StyleSheet, TouchableOpacity } from 'react-native';
-import { Text, View } from '../components/Themed';
+import { Text, View } from '../../components/Themed';
 import React from 'react';
 import {  Button } from 'react-native';
 import * as Haptics from 'expo-haptics';
 import * as NativeStack from '@react-navigation/native-stack';
-import { RootStackParamList } from '../types';
+import { RootStackParamList } from '../../types';
 import { useState } from 'react';
 import { setStatusBarNetworkActivityIndicatorVisible } from 'expo-status-bar';
 import { Image ,ImageBackground} from 'react-native';
 
-export default function History({ navigation }: NativeStack.NativeStackScreenProps<RootStackParamList, 'History'>) {
+export default function StressSelect({ navigation }: NativeStack.NativeStackScreenProps<RootStackParamList, 'StressSelect'>) {
   return (
     <View style={styles.container}>
       <Button color = "green" title='ホーム画面へ' onPress={() => navigation.goBack()} />
